requirements.txt
asgiref==3.7.2
attrs==23.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
Django==5.0.1
django-ratelimit==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.4.0
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
firebase-admin==6.4.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
msgpack==1.0.7
openapi-codec==1.3.2
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-flutterwave==1.2.2
pytz==2023.3.post1
PyYAML==6.0.1
ratelimit==2.2.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
setuptools==69.1.1
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
